Add SqlServerStatsCollector
