+++ plugin/trino-sqlserver/src/main/java/io/trino/plugin/sqlserver/SqlServerClient.java
@@ -79,8 +79,6 @@ import io.trino.spi.expression.ConnectorExpression;
 import io.trino.spi.predicate.Domain;
 import io.trino.spi.predicate.Range;
 import io.trino.spi.predicate.ValueSet;
-import io.trino.spi.statistics.ColumnStatistics;
-import io.trino.spi.statistics.Estimate;
 import io.trino.spi.statistics.TableStatistics;
 import io.trino.spi.type.CharType;
 import io.trino.spi.type.DecimalType;
@@ -115,7 +113,6 @@ import java.time.ZoneId;
 import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeFormatterBuilder;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -125,9 +122,7 @@ import java.util.stream.Stream;
 
 import static com.google.common.base.MoreObjects.firstNonNull;
 import static com.google.common.base.Preconditions.checkArgument;
-import static com.google.common.base.Preconditions.checkState;
 import static com.google.common.base.Throwables.throwIfInstanceOf;
-import static com.google.common.collect.MoreCollectors.toOptional;
 import static com.google.common.util.concurrent.MoreExecutors.directExecutor;
 import static io.airlift.slice.Slices.wrappedBuffer;
 import static io.trino.plugin.jdbc.CaseSensitivity.CASE_INSENSITIVE;
@@ -229,6 +224,7 @@ public class SqlServerClient
 
     private final ConnectorExpressionRewriter<ParameterizedExpression> connectorExpressionRewriter;
     private final AggregateFunctionRewriter<JdbcExpression, ?> aggregateFunctionRewriter;
+    private final SqlServerStatsCollector statsCollector;
 
     private static final int MAX_SUPPORTED_TEMPORAL_PRECISION = 7;
 
@@ -293,6 +289,7 @@ public class SqlServerClient
             ConnectionFactory connectionFactory,
             QueryBuilder queryBuilder,
             IdentifierMapping identifierMapping,
+            SqlServerStatsCollector statsCollector,
             RemoteQueryModifier queryModifier)
     {
         super("\"", connectionFactory, queryBuilder, config.getJdbcTypesMappedToVarchar(), identifierMapping, queryModifier, true);
@@ -341,6 +338,8 @@ public class SqlServerClient
                         .add(new ImplementSqlServerVariancePop())
                         // SQL Server doesn't have covar_samp and covar_pop functions so we can't implement pushdown for them
                         .build());
+
+        this.statsCollector = requireNonNull(statsCollector, "statsCollector is null");
     }
 
     @Override
@@ -717,7 +716,7 @@ public class SqlServerClient
             return TableStatistics.empty();
         }
         try {
-            return readTableStatistics(session, handle);
+            return statsCollector.readTableStatistics(session, handle, this);
         }
         catch (SQLException | RuntimeException e) {
             throwIfInstanceOf(e, TrinoException.class);
@@ -746,128 +745,6 @@ public class SqlServerClient
         catch (SQLException e) {
             throw new TrinoException(JDBC_ERROR, "Failed to get table handle for procedure query. " + firstNonNull(e.getMessage(), e), e);
         }
-    }
-
-    private TableStatistics readTableStatistics(ConnectorSession session, JdbcTableHandle table)
-            throws SQLException
-    {
-        checkArgument(table.isNamedRelation(), "Relation is not a table: %s", table);
-
-        try (Connection connection = connectionFactory.openConnection(session);
-                Handle handle = Jdbi.open(connection)) {
-            RemoteTableName remoteTableName = table.getRequiredNamedRelation().getRemoteTableName();
-            String catalog = remoteTableName.getCatalogName().orElse(null);
-            String schema = remoteTableName.getSchemaName().orElse(null);
-            String tableName = remoteTableName.getTableName();
-
-            StatisticsDao statisticsDao = new StatisticsDao(handle);
-            Long tableObjectId = statisticsDao.getTableObjectId(catalog, schema, tableName);
-            if (tableObjectId == null) {
-                // Table not found
-                return TableStatistics.empty();
-            }
-
-            Long rowCount = statisticsDao.getRowCount(tableObjectId);
-            if (rowCount == null) {
-                // Table disappeared
-                return TableStatistics.empty();
-            }
-
-            if (rowCount == 0) {
-                return TableStatistics.empty();
-            }
-
-            TableStatistics.Builder tableStatistics = TableStatistics.builder();
-            tableStatistics.setRowCount(Estimate.of(rowCount));
-
-            Map<String, String> columnNameToStatisticsName = getColumnNameToStatisticsName(table, statisticsDao, tableObjectId);
-
-            for (JdbcColumnHandle column : this.getColumns(session, table)) {
-                String statisticName = columnNameToStatisticsName.get(column.getColumnName());
-                if (statisticName == null) {
-                    // No statistic for column
-                    continue;
-                }
-
-                double averageColumnLength;
-                long notNullValues = 0;
-                long nullValues = 0;
-                long distinctValues = 0;
-
-                try (CallableStatement showStatistics = handle.getConnection().prepareCall("DBCC SHOW_STATISTICS (?, ?)")) {
-                    showStatistics.setString(1, format("%s.%s.%s", catalog, schema, tableName));
-                    showStatistics.setString(2, statisticName);
-
-                    boolean isResultSet = showStatistics.execute();
-                    checkState(isResultSet, "Expected SHOW_STATISTICS to return a result set");
-                    try (ResultSet resultSet = showStatistics.getResultSet()) {
-                        checkState(resultSet.next(), "No rows in result set");
-
-                        averageColumnLength = resultSet.getDouble("Average Key Length"); // NULL values are accounted for with length 0
-
-                        checkState(!resultSet.next(), "More than one row in result set");
-                    }
-
-                    isResultSet = showStatistics.getMoreResults();
-                    checkState(isResultSet, "Expected SHOW_STATISTICS to return second result set");
-                    showStatistics.getResultSet().close();
-
-                    isResultSet = showStatistics.getMoreResults();
-                    checkState(isResultSet, "Expected SHOW_STATISTICS to return third result set");
-                    try (ResultSet resultSet = showStatistics.getResultSet()) {
-                        while (resultSet.next()) {
-                            resultSet.getObject("RANGE_HI_KEY");
-                            if (resultSet.wasNull()) {
-                                // Null fraction
-                                checkState(resultSet.getLong("RANGE_ROWS") == 0, "Unexpected RANGE_ROWS for null fraction");
-                                checkState(resultSet.getLong("DISTINCT_RANGE_ROWS") == 0, "Unexpected DISTINCT_RANGE_ROWS for null fraction");
-                                checkState(nullValues == 0, "Multiple null fraction entries");
-                                nullValues += resultSet.getLong("EQ_ROWS");
-                            }
-                            else {
-                                // TODO discover min/max from resultSet.getXxx("RANGE_HI_KEY")
-                                notNullValues += resultSet.getLong("RANGE_ROWS") // rows strictly within a bucket
-                                        + resultSet.getLong("EQ_ROWS"); // rows equal to RANGE_HI_KEY
-                                distinctValues += resultSet.getLong("DISTINCT_RANGE_ROWS") // NDV strictly within a bucket
-                                        + (resultSet.getLong("EQ_ROWS") > 0 ? 1 : 0);
-                            }
-                        }
-                    }
-                }
-
-                ColumnStatistics statistics = ColumnStatistics.builder()
-                        .setNullsFraction(Estimate.of(
-                                (notNullValues + nullValues == 0)
-                                        ? 1
-                                        : (1.0 * nullValues / (notNullValues + nullValues))))
-                        .setDistinctValuesCount(Estimate.of(distinctValues))
-                        .setDataSize(Estimate.of(rowCount * averageColumnLength))
-                        .build();
-
-                tableStatistics.setColumnStatistics(column, statistics);
-            }
-
-            return tableStatistics.build();
-        }
-    }
-
-    private static Map<String, String> getColumnNameToStatisticsName(JdbcTableHandle table, StatisticsDao statisticsDao, Long tableObjectId)
-    {
-        List<String> singleColumnStatistics = statisticsDao.getSingleColumnStatistics(tableObjectId);
-
-        Map<String, String> columnNameToStatisticsName = new HashMap<>();
-        for (String statisticName : singleColumnStatistics) {
-            String columnName = statisticsDao.getSingleColumnStatisticsColumnName(tableObjectId, statisticName);
-            if (columnName == null) {
-                // Table or statistics disappeared
-                continue;
-            }
-
-            if (columnNameToStatisticsName.putIfAbsent(columnName, statisticName) != null) {
-                log.debug("Multiple statistics for %s in %s: %s and %s", columnName, table, columnNameToStatisticsName.get(columnName), statisticName);
-            }
-        }
-        return columnNameToStatisticsName;
     }
 
     // SQL Server has non-standard LIKE semantics:
@@ -1320,66 +1197,5 @@ public class SqlServerClient
         return Failsafe
                 .with(retryPolicy)
                 .get(supplier);
-    }
-
-    private static class StatisticsDao
-    {
-        private final Handle handle;
-
-        public StatisticsDao(Handle handle)
-        {
-            this.handle = requireNonNull(handle, "handle is null");
-        }
-
-        Long getTableObjectId(String catalog, String schema, String tableName)
-        {
-            return handle.createQuery("SELECT object_id(:table)")
-                    .bind("table", format("%s.%s.%s", catalog, schema, tableName))
-                    .mapTo(Long.class)
-                    .one();
-        }
-
-        Long getRowCount(long tableObjectId)
-        {
-            return handle.createQuery("" +
-                            "SELECT sum(rows) row_count " +
-                            "FROM sys.partitions " +
-                            "WHERE object_id = :object_id " +
-                            "AND index_id IN (0, 1)") // 0 = heap, 1 = clustered index, 2 or greater = non-clustered index
-                    .bind("object_id", tableObjectId)
-                    .mapTo(Long.class)
-                    .one();
-        }
-
-        List<String> getSingleColumnStatistics(long tableObjectId)
-        {
-            return handle.createQuery("" +
-                            "SELECT s.name " +
-                            "FROM sys.stats AS s " +
-                            "JOIN sys.stats_columns AS sc ON s.object_id = sc.object_id AND s.stats_id = sc.stats_id " +
-                            "WHERE s.object_id = :object_id " +
-                            "GROUP BY s.name " +
-                            "HAVING count(*) = 1 " +
-                            "ORDER BY s.name")
-                    .bind("object_id", tableObjectId)
-                    .mapTo(String.class)
-                    .list();
-        }
-
-        String getSingleColumnStatisticsColumnName(long tableObjectId, String statisticsName)
-        {
-            return handle.createQuery("" +
-                            "SELECT c.name " +
-                            "FROM sys.stats AS s " +
-                            "JOIN sys.stats_columns AS sc ON s.object_id = sc.object_id AND s.stats_id = sc.stats_id " +
-                            "JOIN sys.columns AS c ON sc.object_id = c.object_id AND c.column_id = sc.column_id " +
-                            "WHERE s.object_id = :object_id " +
-                            "AND s.name = :statistics_name")
-                    .bind("object_id", tableObjectId)
-                    .bind("statistics_name", statisticsName)
-                    .mapTo(String.class)
-                    .collect(toOptional()) // verify there is no more than 1 column name returned
-                    .orElse(null);
-        }
     }
 }
